pyproject.toml
[project]
name = "sr2"
version = "0.1.0"
description = "Context engineering library for AI agents"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# --- Runtime extras (pick what your agent needs) ---
runtime = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
]
telegram = ["python-telegram-bot>=21.0"]
postgres = ["asyncpg>=0.29"]
a2a = ["a2a-sdk>=0.3"]
mcp = ["mcp>=1.0"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

# --- Convenience bundles ---
all = [
    "sr2[runtime,telegram,postgres,a2a,mcp,otel]",
]

# --- Dev / testing ---
dev = [
    "sr2[all]",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "httpx>=0.27",
]

[project.scripts]
sr2-agent = "runtime.cli:main"
sr2-config-docs = "schema_gen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sr2", "src/runtime", "agents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
